100 - Extract config into separate file
