+++ tel2puml/otel_to_pv/data_sources/json_data_source/json_jq_converter.py
@@ -4,9 +4,7 @@ from typing import Any, Generator
 
 import jq  # type: ignore[import-not-found]
 
-from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
-    FieldSpec,
-)
+from tel2puml.otel_to_pv.config import FieldSpec
 
 
 class JQVariableTree:
