+++ tests/tel2puml/otel_to_pv/data_sources/json_data_source/test_json_datasource.py
@@ -13,9 +13,9 @@ from tel2puml.otel_to_pv.data_sources.json_data_source.json_datasource import (
     JSONDataSource,
 )
 from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
-    IngestDataConfig,
-    OTelEvent,
+    OTelEvent
 )
+from tel2puml.otel_to_pv.config import IngestDataConfig
 
 
 class TestJSONDataSource:
