+++ tests/tel2puml/otel_to_pv/data_sources/json_data_source/test_json_jq_converter.py
@@ -2,9 +2,7 @@
 from typing import Any
 import pytest
 
-from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
-    FieldSpec,
-)
+from tel2puml.otel_to_pv.config import FieldSpec
 from tel2puml.otel_to_pv.data_sources.json_data_source.json_jq_converter \
     import (
         JQVariableTree,
